+++ src/functionalTest/kotlin/io/github/surpsg/CoverageReportTaskTest.kt
@@ -7,7 +7,6 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.MethodSource
 
-
 class CoverageReportTaskTest : BaseOfflinsTest() {
 
     companion object {